refactor: onboarding widget 분리
